task/fuse: More options and failure handling
Handling for cleanup/teardown of a fuse mount and process
in the not-so-uncommon case where something has gone horribly wrong.

Option 'cleanup_on_failure' can be set to true, allowing
proper cleanup if the fuse process is expected to fail
for debugging purposes.

Option 'timeout' can be set to a number of seconds, after which
waiting for the ceph-fuse process to exit will cancel, and the process
will be killed.  The wait period starts at unmount.

Signed-off-by: Sam Lang <sam.lang@inktank.com>